requirements.txt
asgiref==3.7.2
crispy-tailwind==1.0.3
Django==5.0.3
django-crispy-forms==2.3
django-environ==0.11.2
django-storages==1.14.4
Pillow==9.4.0
python-decouple==3.8
python-utils==3.8.2
pytz==2023.2
sqlparse==0.4.4
typing_extensions==4.12.2
